// types.ts
export interface TimeZone {
  name: string;
  identifier: string;
}

export interface WorldTimeAPIResponse {
  abbreviation: string;
  client_ip: string;
  datetime: string;
  day_of_week: number;
  day_of_year: number;
  dst: boolean;
  dst_from: string | null;
  dst_offset: number;
  dst_until: string | null;
  raw_offset: number;
  timezone: string;
  unixtime: number;
  utc_datetime: string;
  utc_offset: string;
  week_number: number;
}

export interface ConversionResult {
  cityName: string;
  time: string;
  timeDiff: string;
  date: string;
}

export type TimeFormat = '12h' | '24h';

export type ModalType = 'About' | 'Contact' | 'Guide' | 'Privacy Policy' | 'Terms of Service' | 'DMCA';